Fix useUSDCPrice bug / V5.8 (#41)
* Add USDC price display

This commit adds the USDC Price to the advanced swap details when a user initiates a swap. Shows whatever the Input token is in USDC. Also correct about page info.

* Correct Token analytics info

This commit allows for the FISH token contract stats to be displayed when the user clicks the 'fish' icon. Page no longer refreshes after user clicks icon, and errors are no longer present.

* UI Corrections

Correct text color the the "Analytics" links in the 🐟info / analytics modal

Corrects "Pending" transaction text in light mode to be visible

* Correct USDC amount info

This commit corrects the info in the Advanced Swap Details to be more specific about the USDC price of token A

* Merge with PP DEV master

* Update Swap details

- Adds USDC trade amounts for both input and output tokens
- Updates advaced swap details:
  - rate, inverse rate, USD rate, also adds USD amount of 'minimum
  received'

* Correct manageToken scroll issue

- Adds `overflow: scroll` to TokenList Wrapper CSS
- Add `position: bottom` to the footer in the ManageTokens
  token list modal

V5.7

* Corrects bug found w/ useUSDCPrice functionality

* Roulette
- Sets [otherField]: { currencyId: currencyId } in
src/state/swap/reducer.ts

V1.2

Co-authored-by: David E. Felton <hiturunk@gmail.com>